Rename platform files and submodules
