+++ .gitmodules
@@ -7,8 +7,8 @@
 [submodule "FreeRTOS-Plus/Source/FreeRTOS-Plus-TCP"]
 	path = FreeRTOS-Plus/Source/FreeRTOS-Plus-TCP
 	url = https://github.com/FreeRTOS/FreeRTOS-Plus-TCP
-[submodule "FreeRTOS-Plus/Source/AWS/device-shadow-for-aws-iot-embedded-sdk"]
-	path = FreeRTOS-Plus/Source/AWS/device-shadow-for-aws-iot-embedded-sdk
+[submodule "FreeRTOS-Plus/Source/AWS/device-shadow"]
+	path = FreeRTOS-Plus/Source/AWS/device-shadow
 	url = https://github.com/aws/device-shadow-for-aws-iot-embedded-sdk
 [submodule "FreeRTOS-Plus/Source/Application-Protocols/coreMQTT"]
 	path = FreeRTOS-Plus/Source/Application-Protocols/coreMQTT
