+++ FreeRTOS-Plus/Demo/AWS/Device_Shadow_Windows_Simulator/Device_Shadow_Demo/DemoTasks/shadow_demo_helpers.c
@@ -53,7 +53,7 @@
 #include "retry_utils.h"
 
 /* Transport interface implementation include header for TLS. */
-#include "tls_freertos.h"
+#include "freertos_plus_tcp_mbedtls.h"
 
 /*------------- Demo configurations -------------------------*/
