+++ FreeRTOS-Plus/Demo/AWS/Device_Shadow_Windows_Simulator/Device_Shadow_Demo/WIN32.vcxproj.filters
@@ -59,16 +59,16 @@
     <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos">
       <UniqueIdentifier>{fcf93295-15e2-4a84-a5e9-b3c162e9f061}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls">
       <UniqueIdentifier>{8a0aa896-6b3a-49b3-997e-681f0d1949ae}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport">
       <UniqueIdentifier>{c5a01679-3e7a-4320-97ac-ee5b872c1650}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include">
       <UniqueIdentifier>{c992824d-4198-46b2-8d59-5f99ab9946ab}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\src">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport">
       <UniqueIdentifier>{6a35782c-bc09-42d5-a850-98bcb668a4dc}</UniqueIdentifier>
     </Filter>
     <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\standard\coreJSON">
@@ -77,7 +77,7 @@
     <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\AWS">
       <UniqueIdentifier>{0dacb84e-5cc3-4eed-8fb1-68b6e4741f77}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\AWS\device-shadow-for-aws-iot-embedded-sdk">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\AWS\device-shadow">
       <UniqueIdentifier>{3dc3c37b-6417-4a84-a16e-b1acf21c04cf}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
@@ -150,20 +150,20 @@
     <ClCompile Include="..\..\..\..\Source\Application-Protocols\coreMQTT\source\core_mqtt.c">
       <Filter>FreeRTOS+\FreeRTOS IoT Libraries\standard\coreMQTT\src</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_freertos_port.c">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\mbedtls\mbedtls_freertos_port.c">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_error.c">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\mbedtls\mbedtls_error.c">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\freertos\retry_utils\retry_utils_freertos.c">
+    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\retry_utils\retry_utils.c">
       <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\freertos\transport\src\tls_freertos.c">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\src</Filter>
+    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\transport\tls\freertos_plus_tcp_mbedtls.c">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\freertos\transport\src\freertos_sockets_wrapper.c">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\src</Filter>
+    <ClCompile Include="..\..\..\..\Source\Application-Protocols\platform\transport\freertos_plus_tcp_sockets_wrapper.c">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport</Filter>
     </ClCompile>
     <ClCompile Include="..\..\..\..\ThirdParty\mbedtls\library\aes.c" >
       <Filter>FreeRTOS+\mbedtls\library</Filter>
@@ -412,8 +412,8 @@
     <ClCompile Include="..\..\..\..\Source\coreJSON\source\core_json.c">
       <Filter>FreeRTOS+\FreeRTOS IoT Libraries\standard\coreJSON</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\..\Source\AWS\device-shadow-for-aws-iot-embedded-sdk\source\shadow.c">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\AWS\device-shadow-for-aws-iot-embedded-sdk</Filter>
+    <ClCompile Include="..\..\..\..\Source\AWS\device-shadow\source\shadow.c">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\AWS\device-shadow</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -506,17 +506,17 @@
       <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform</Filter>
     </ClInclude>
     <ClInclude Include="mbedtls_config.h" />
-    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_error.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\mbedtls\mbedtls_error.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\threading_alt.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\mbedtls\threading_alt.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\freertos\transport\include\freertos_sockets_wrapper.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include</Filter>
+    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\transport\include\freertos_plus_tcp_sockets_wrapper.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\freertos\transport\include\tls_freertos.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include</Filter>
+    <ClInclude Include="..\..\..\..\Source\Application-Protocols\platform\transport\include\freertos_plus_tcp_mbedtls.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\..\ThirdParty\mbedtls\include\mbedtls\aes.h" />
     <ClInclude Include="..\..\..\..\ThirdParty\mbedtls\include\mbedtls\aesni.h" />
