+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Basic_TLS/DemoTasks/BasicTLSMQTTExample.c
@@ -57,7 +57,7 @@
 #include "retry_utils.h"
 
 /* Transport interface implementation include header for TLS. */
-#include "tls_freertos.h"
+#include "freertos_plus_tcp_mbedtls.h"
 
 /*-----------------------------------------------------------*/
