+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Keep_Alive/DemoTasks/KeepAliveMQTTExample.c
@@ -60,7 +60,7 @@
 #include "retry_utils.h"
 
 /* Transport interface include. */
-#include "plaintext_freertos.h"
+#include "freertos_plus_tcp_plaintext.h"
 
 /*-----------------------------------------------------------*/
