+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Mutual_Auth/DemoTasks/MutualAuthMQTTExample.c
@@ -58,7 +58,7 @@
 #include "retry_utils.h"
 
 /* Transport interface implementation include header for TLS. */
-#include "tls_freertos.h"
+#include "freertos_plus_tcp_mbedtls.h"
 
 /*-----------------------------------------------------------*/
