+++ FreeRTOS-Plus/Demo/coreMQTT_Windows_Simulator/MQTT_Plain_Text/DemoTasks/PlaintextMQTTExample.c
@@ -56,7 +56,7 @@
 #include "retry_utils.h"
 
 /* Transport interface include. */
-#include "plaintext_freertos.h"
+#include "freertos_plus_tcp_plaintext.h"
 
 /*-----------------------------------------------------------*/
