+++ FreeRTOS-Plus/Demo/corePKCS11_MQTT_Mutual_Auth_Windows_Simulator/WIN32.vcxproj.filters
@@ -59,16 +59,16 @@
     <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos">
       <UniqueIdentifier>{fcf93295-15e2-4a84-a5e9-b3c162e9f061}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls">
       <UniqueIdentifier>{8a0aa896-6b3a-49b3-997e-681f0d1949ae}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport">
       <UniqueIdentifier>{c5a01679-3e7a-4320-97ac-ee5b872c1650}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include">
       <UniqueIdentifier>{c992824d-4198-46b2-8d59-5f99ab9946ab}</UniqueIdentifier>
     </Filter>
-    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\src">
+    <Filter Include="FreeRTOS+\FreeRTOS IoT Libraries\platform\transport">
       <UniqueIdentifier>{6a35782c-bc09-42d5-a850-98bcb668a4dc}</UniqueIdentifier>
     </Filter>
     <Filter Include="ThirdParty">
@@ -166,16 +166,16 @@
     <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\coreMQTT\source\core_mqtt_state.c">
       <Filter>coreMQTT</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\freertos\transport\src\freertos_sockets_wrapper.c">
+    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\transport\freertos_plus_tcp_sockets_wrapper.c">
       <Filter>Transport</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_error.c">
+    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\mbedtls\mbedtls_error.c">
       <Filter>ThirdParty</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_freertos_port.c">
+    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\mbedtls\mbedtls_freertos_port.c">
       <Filter>Transport</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\freertos\retry_utils\retry_utils_freertos.c">
+    <ClCompile Include="..\..\..\FreeRTOS-Plus\Source\Application-Protocols\platform\retry_utils\retry_utils.c">
       <Filter>Transport</Filter>
     </ClCompile>
     <ClCompile Include="main.c" />
@@ -452,7 +452,7 @@
     <ClCompile Include="..\..\ThirdParty\mbedtls\library\xtea.c">
       <Filter>ThirdParty\mbedtls</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Source\Application-Protocols\platform\freertos\transport\src\tls_freertos_pkcs11.c">
+    <ClCompile Include="..\..\Source\Application-Protocols\platform\transport\tls\freertos_plus_tcp_mbedtls_pkcs11.c">
       <Filter>Transport</Filter>
     </ClCompile>
   </ItemGroup>
@@ -545,17 +545,17 @@
       <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform</Filter>
     </ClInclude>
     <ClInclude Include="mbedtls_config.h" />
-    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\mbedtls_error.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\mbedtls\mbedtls_error.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\freertos\mbedtls\threading_alt.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\mbedtls</Filter>
+    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\mbedtls\threading_alt.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\mbedtls</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\freertos\transport\include\freertos_sockets_wrapper.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include</Filter>
+    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\transport\include\freertos_plus_tcp_sockets_wrapper.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\freertos\transport\include\tls_freertos.h">
-      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\freertos\transport\include</Filter>
+    <ClInclude Include="..\..\..\Source\Application-Protocols\platform\transport\include\freertos_plus_tcp_mbedtls.h">
+      <Filter>FreeRTOS+\FreeRTOS IoT Libraries\platform\transport\include</Filter>
     </ClInclude>
     <ClInclude Include="core_mqtt_config.h">
       <Filter>coreMQTT</Filter>
