+++ FreeRTOS-Plus/Source/Application-Protocols/platform/transport/tls/freertos_plus_tcp_mbedtls.c
@@ -36,10 +36,10 @@
 #include "FreeRTOS_Sockets.h"
 
 /* TLS transport header. */
-#include "tls_freertos.h"
+#include "freertos_plus_tcp_mbedtls.h"
 
 /* FreeRTOS Socket wrapper include. */
-#include "freertos_sockets_wrapper.h"
+#include "freertos_plus_tcp_sockets_wrapper.h"
 
 /* mbedTLS util includes. */
 #include "mbedtls_error.h"
